requirements.txt
django==5.1.5
requests==2.32.3
pandas==2.2.3
matplotlib==3.10.0
mysqlclient==2.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0